changes to some data types and inclusion of openGA

+++ include/simulation/EdgeTrajectory.hpp
@@ -12,11 +12,11 @@ struct TrainState {
   /**
    * Single train state
    */
-  u_int64_t timestep;    // [0, n_timesteps - 1]
-  size_t    edge;        // [0, network.edges.size() - 1]
-  double    position;    // [0, 1]
-  bool      orientation; // true, false = forward, backward
-  double    speed;       // (-Inf, Inf)
+  size_t timestep;    // [0, n_timesteps - 1]
+  size_t edge;        // [0, network.edges.size() - 1]
+  double position;    // [0, 1]
+  bool   orientation; // true, false = forward, backward
+  double speed;       // (-Inf, Inf)
 };
 
 enum EdgeEntryOutcome {
@@ -35,13 +35,13 @@ struct EdgeEntry {
 };
 
 struct EdgeTraversal {
-  u_int64_t from_timestep;
-  size_t    from_edge;
-  bool      from_exit_point;      // true, false = forward, backward
-  size_t    vertex;               // [0, network.edges.size() - 1]
-  bool      crossing_orientation; // true, false = forward, backward
-  double    leftover_movement;    // (0, Inf)
-  double    speed;                // (-Inf, Inf)
+  size_t from_timestep;
+  size_t from_edge;
+  bool   from_exit_point;      // true, false = forward, backward
+  size_t vertex;               // [0, network.edges.size() - 1]
+  bool   crossing_orientation; // true, false = forward, backward
+  double leftover_movement;    // (0, Inf)
+  double speed;                // (-Inf, Inf)
 };
 
 class EdgeTrajectory {
@@ -51,10 +51,10 @@ class EdgeTrajectory {
   std::reference_wrapper<const SimulationInstance> instance_r;
   std::reference_wrapper<const Train>              train_r;
 
-  u_int64_t first_timestep; // [0, n_timesteps - 1]
-  u_int64_t last_timestep;  // [0, n_timesteps - 1]
-  size_t    edge;           // [0, network.edges.1ize() - 1]
-  bool      orientation;    // true, false = forward, backward
+  size_t first_timestep; // [0, n_timesteps - 1]
+  size_t last_timestep;  // [0, n_timesteps - 1]
+  size_t edge;           // [0, network.edges.1ize() - 1]
+  bool   orientation;    // true, false = forward, backward
 
   std::vector<double> positions; // [0, 1]
   std::vector<double> speeds;    // (-Inf, Inf)
@@ -73,8 +73,8 @@ public:
 
   cda_rail::ScheduledStop get_stop() const;
 
-  u_int64_t                    get_first_timestep() const;
-  u_int64_t                    get_last_timestep() const;
+  size_t                       get_first_timestep() const;
+  size_t                       get_last_timestep() const;
   size_t                       get_edge() const;
   bool                         get_orientation() const;
   const std::vector<double>&   get_positions() const;
